requirements.txt
beautifulsoup4==4.6.0
evernote==1.25.3
python-dotenv==0.18.0
